Moves League general methods to List
