+++ lib/stat_tracker.rb
@@ -1,9 +1,11 @@
 require "./lib/league_stats"
 require './lib/team_stats'
+require "./lib/list_mod"
 
 class StatTracker
   include TeamStats
   include LeagueStats
+  include List
 
   attr_accessor :games, :game_teams, :teams
